requirements.txt
qtpy~=2.4.2
pymodaq~=4.4.7
numpy~=1.26.4
pytango~=10.0.0